Fix adjacent vertex calculation
Fix a bug that did not cycle through adjacent vertices correctly. The
problem was that seed edge was selected randomly. Since the adjacent
vertices are determined by cycling around the vertex along faces, the
procedure stopped early if no adjacent face was found. Now, the seed edge
is selected s. t. it is a border edge if existing.

This means in particular that meshes with a vertex adjacent to two faces
that share no edge cannot be handled. The README.MD was updated
accordingly.

Some minor code-style fixes are also included.